feat(compiler): Intricacies of startup and restart
