+++ module/compiler/support/server/client.ts
@@ -17,6 +17,8 @@ type FetchEventsConfig<T> = {
   enforceIteration?: boolean;
 };
 
+const SHUTDOWN_TIMEOUT = 3000;
+
 /**
  * Compiler Client Operations
  */
@@ -54,8 +56,29 @@ export class CompilerClient {
   }
 
   /** Stop server */
-  stop(): Promise<boolean> {
+  async stop(): Promise<boolean> {
     return fetch(`${this.#url}/stop`).then(v => v.ok, () => false);
+  }
+
+  /** Stop server and wait for shutdown */
+  async stopAndWait(): Promise<boolean> {
+    const info = await this.info();
+    if (!info) {
+      return false;
+    }
+    await this.stop(); // Start request
+    const start = Date.now();
+    for (; ;) { // Ensure its done
+      try {
+        process.kill(info.compilerPid, 0); // See if process is still running
+      } catch {
+        return true; // If not, its done
+      }
+      await timers.setTimeout(100);
+      if ((Date.now() - start) > SHUTDOWN_TIMEOUT) { // If we exceed the max timeout
+        process.kill(info.compilerPid); // Force kill
+      }
+    }
   }
 
   /** Fetch compiler events */
